moved codec entities into http package
